# tickets_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tickets",
    "version": "0.1.0"
  },
  "paths": {
...
    "/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Register",
        "operationId": "register_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/screenings/{screening_id}/availability/stream": {
      "get": {
        "summary": "Stream Seat Availability",
        "operationId": "stream_seat_availability_screenings__screening_id__availability_stream_get",
        "parameters": [
          {
            "name": "screening_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Screening Id"
            }
          },
          {
            "name": "last_event_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "$",
              "title": "Last Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/screenings/{screening_id}/seats": {
      "get": {
        "summary": "View Selected Seats",
        "operationId": "view_selected_seats_screenings__screening_id__seats_get",
        "parameters": [
          {
            "name": "screening_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Screening Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response View Selected Seats Screenings  Screening Id  Seats Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/screenings/{screening_id}/seat-map": {
      "get": {
        "summary": "View Screening Seat Map",
        "operationId": "view_screening_seat_map_screenings__screening_id__seat_map_get",
        "parameters": [
          {
            "name": "screening_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Screening Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScreeningSeatResponse"
                  },
                  "title": "Response View Screening Seat Map Screenings  Screening Id  Seat Map Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/screenings/{screening_id}/seats/{seat_id}/hold": {
      "post": {
        "summary": "Hold Seat",
        "operationId": "hold_seat_screenings__screening_id__seats__seat_id__hold_post",
        "parameters": [
          {
            "name": "screening_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Screening Id"
            }
          },
          {
            "name": "seat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Seat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "title": "Response Hold Seat Screenings  Screening Id  Seats  Seat Id  Hold Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Release Held Seat",
        "operationId": "release_held_seat_screenings__screening_id__seats__seat_id__hold_delete",
        "parameters": [
          {
            "name": "screening_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Screening Id"
            }
          },
          {
            "name": "seat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Seat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "title": "Response Release Held Seat Screenings  Screening Id  Seats  Seat Id  Hold Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/screenings/{screening_id}/seats/checkout": {
      "post": {
        "summary": "Checkout Seats",
        "operationId": "checkout_seats_screenings__screening_id__seats_checkout_post",
        "parameters": [
          {
            "name": "screening_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Screening Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "title": "Response Checkout Seats Screenings  Screening Id  Seats Checkout Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/screenings/{screening_id}/checkout/": {
      "get": {
        "summary": "Get Checkout",
        "operationId": "get_checkout_screenings__screening_id__checkout__get",
        "parameters": [
          {
            "name": "screening_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Screening Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Checkout Screenings  Screening Id  Checkout  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Cancel Checkout",
        "operationId": "cancel_checkout_screenings__screening_id__checkout__delete",
        "parameters": [
          {
            "name": "screening_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Screening Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "title": "Response Cancel Checkout Screenings  Screening Id  Checkout  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/purchases/{ticket_id}": {
      "get": {
        "summary": "Get Ticket",
        "operationId": "get_ticket_purchases__ticket_id__get",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "migrated_seat_count": {
            "type": "integer",
            "title": "Migrated Seat Count"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email",
          "migrated_seat_count"
        ],
        "title": "LoginResponse"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x_internal_service_token"
      }
    }
  }
}